+++ SampleApplication/SampleApplication/NetworkLayer/NetworkLayer.swift
@@ -29,4 +29,30 @@ struct NetworkLayer {
     
     return try decoder.decode(PaymentSession.self, from: data)
   }
+  
+  func submitPaymentSession(paymentSessionId: String,
+                            request: SubmitPaymentSessionRequest) async throws -> CheckoutComponents.PaymentSessionSubmissionResult {
+    let encoder = JSONEncoder()
+    encoder.keyEncodingStrategy = .convertToSnakeCase
+    let requestBody = try encoder.encode(request)
+    
+    // Don't send requests to this API but have a wrapper on your backend to keep your private key safe.
+    // Otherwise you would have your private key bundled in your application and get it leaked.
+    let url = URL(string: "https://api.sandbox.checkout.com/payment-sessions/\(paymentSessionId)/submit")!
+    
+    var request = URLRequest(url: url)
+    request.httpBody = requestBody
+    request.httpMethod = "POST"
+    request.addValue("Bearer " + EnvironmentVars.secretKey, forHTTPHeaderField: "Authorization")
+    
+    let (data, response) = try await URLSession.shared.data(for: request)
+    if let response = response as? HTTPURLResponse {
+      print(response.statusCode)
+    }
+    let decoder = JSONDecoder()
+    decoder.keyDecodingStrategy = .convertFromSnakeCase
+    
+    return try decoder.decode(CheckoutComponents.PaymentSessionSubmissionResult.self,
+                              from: data)
+  }
 }